gateway: return single object when listing without a prefix
please see discussion on
https://storjlabs.atlassian.net/browse/USR-537

as laid out in code comments also:

The most S3-compatible thing we could do here is ask
the satellite to list all siblings of this prefix that
share the same parent encryption key, decrypt all of
them, then only return the ones that have this same
unencrypted prefix.

This is terrible from a performance perspective, and
this remains a pretty big architectural challenge for
us. Bummer!

Luckily, it turns out, many of the usages of listing
without a /-suffix are simply to provide a sort of
StatObject like feature. In fact, for example,
Duplicity never calls list without a /-suffix in a
case where it expects to get back more than one
result. Hooray!

So, we could either

 1) return an error here, guaranteeing nothing works.
 2) do the full S3 compatible thing, which has terrible
    performance in a really common case (StatObject-like
    functionality, e.g., apps that have a specific key
    they simply want to know the size for)
 3) handle strictly more of the use cases than #1 does
    without loss of performance by turning this into a
    StatObject when a /-suffix is missing.

In terms of Venn diagrams, #3 (this changeset) breaks
nothing that isn't already broken, while fixing the
behavior in many common use cases. It doesn't solve all
problems of course, but it solves more of them.

Addendum: we also return a prefix entry if the object
is a /-delineated prefix, in case callers are trying
to do essentially a StatPrefix-type call and confirm
a prefix exists

Change-Id: Ic56e9d1ca6e929f8bc22ac724ace5b705353089c