Multitenant Gateway
This change adds support for multitenancy. The S3 Gateway can accept any
Tardigrade Access Grant as the S3 Access Key and request the Tardigrade
Satellites with it.

As a consequence all gateway flags related to configuring access key,
secret key, and access grant are removed.

We cannot support static websites anymore, so the respective code and
tests are removed too.

We still have to figure out how to support user agents for value
attribution, but this will be addressed in a future change.

Change-Id: I32210fedf4e0995f29d5bd296554695fc8c013f4