testsuite: improve stability of TestUploadDownload
Sometimes the test fails because it is requesting the gateway before the
backend implementation of the gateway is fully initialized. The test
waits for the gateway process to start and even tries to connect to its
port, but it seems this not enough. Sometimes requests fails with 503
error code, which means that the gateway process is started and
listening, but the backend has not been initialized yet.

The solution is: instead of just trying to connect to the gateway port,
try to list the buckets. Successfully listing the buckets ensures that
the backend has been initialized.

Change-Id: I7ba7cb4d639d1b4782b4283b17786cffa1dc28d4