main.go: User better spelling in error message
Use a better spelling for the compound word "non interactive"; according
some sources it should be "noninteractive", "non-interactive" is also
valid. It looks that the current tendency is not to use dash in this
specific case but because the command line flag uses dash, we keep it
stuck with it.

Source:
https: //www.chicagomanualofstyle.org/qanda/data/faq/topics/HyphensEnDashesEmDashes/faq0079.html

Change-Id: Iabc4a31051eca5b0882ae42cf75b69f51db8840c