Remove --client.user-agent flag
The multi-tenant gateway does not support the --client.user-agent flag
as the gateway is not mapped to a single Tardigrade project anymore.

In future, the gateway will take the User Agent from the incoming S3
request and pass it to the uplink request to the Tardigrade satellite.

For now, we are removing the flag to avoid the confusion.

Change-Id: I3f1db93d7ec7888cf8daaa84aa939e3a3dade23f