miniogw: allow last-byte-pos to be greater than content length
According to the definition for Byte Ranges in the HTTP specification:
"If the last-byte-pos value is absent, or if the value is greater than
or equal to the current length of the entity-body, last-byte-pos is
taken to be equal to one less than the current length of the entity-
body in bytes."

This is now done internally by the Uplink library, but we still have a
check against the "allow last-byte-pos to be greater than content
length" in the Gateway code.

This fix removes this check and updates the test cases.

Change-Id: Icdb3ded3cf51d54612c3c51a7c3ead3983a5c9c0