miniogw: make ListObjects(V2) fully respect limits
ListObjects/ListObjectsV2 and listSingleObject/listSingleObjectV2 (code
path for stat in ListObjects[V2]]) did not fully respect MaxKeys limit.
One of the main problems was that if the bucket contains many prefixes
on a non-recursive listing, ListObjects(V2) won't return the next token
to continue listing, though indicating more items to list. Listing items
with a prefix that don't terminate with delimiter could end with listing
too many items ("stat" code path). This change fixes all these issues.
Tests were added to testsuite that test limits handling to avoid
regression in the future.

This change accounts for a bug in libuplink where non-recursively
listing items with a prefix starting from a prefix could result in
getting the cursor as the first item in the list. A temporary workaround
has been introduced to avoid that.

Change-Id: I23390cf988cede20bcf4249ea90c4b5dc5cbdc60