testsuite/server: add TestPutObjectPartSegmentSize
This change adds a test that verifies there won't be an error while
committing a multipart upload with a part size of segment size.

The maximum segment size has been changed to 640 KiB for all tests,
which should be enough for them and less burdensome than the 64 MiB
threshold for segment size-related tests.

Change-Id: Ifdbfd763f45457b0b6419da19fcfd3ecad892c5a