main.go: make access error messages a little friendlier
Currently the access grant is being checked for correctness, and
if it looks correct, a call is made to the Satellite to make sure
the connection is working. The general "Failed to contact Satellite..."
error can be misleading if the error is simply that the access
grant is invalid, so this change produces a slightly more useful
error to the user. As a result of changing this, the big and not
so useful backtrace is removed when running with the flag --releases
in dev mode.

Closes #40

Change-Id: I9d71ecea2b228a5d757534354e6efbf98250c380