ci: don't call check-downgrades from testsuite
Calling check-downgrades from testsuite makes it confuse the top
module's go.mod with the nested module's one. This happens because
check-downgrades cannot easily know it was called from within the nested
module and does not change the directory when it performs the check
against the previous commit's directory. Moreover, this is redundant to
call check-downgrades from the nested module as it can find all the
nested modules when called from the top module.

Change-Id: Ic91b54aea011e9a595bb67ed5e22e72e9175cf6b