miniogw: allow CopyObject to update metadata for same src and dest
Tools like rclone sync use CopyObject endpoint with the same source
and destination to update metadata on an object where only the
modified time has changed, to avoid having to re-upload the exact
same object again from scratch.

Since we've disabled CopyObject on Gateway-MT in production because
server-side copy hasn't been implemented yet, users currently have
to work around rclone hitting "Not implemented" errors by setting
`--no-update-modtime` so it doesn't call this endpoint.

This changes the CopyObject endpoint to update the object's
metadata if source and destination are the same. It's a special
case that works regardless of whether CopyObject has been disabled
or not.

Closes #44

Change-Id: I73737f50d87756ed972ce68149bfaff10c53af48