{miniogw, testsuite/miniogw}: add support for arbitrary prefix/delimiter
This change adds support for arbitrary prefix and delimiter to
ListObjects and ListObjectsV2. The listing code has been rewritten. The
new code chooses one of three ways (fast, non-terminated
prefix-optimized, exhaustive) to list the bucket based on the given
prefix and delimiter.

1. Fast listing

The prefix must be empty or forward slash-terminated, and the delimiter
must be empty or forward slash. Fast listing is how our listing code
worked before.

2. Listing optimized for non-terminated prefixes

The prefix must not end with a forward slash. We had this optimization
before, but it has been revisited. S3-compatible software often checks
if a specific key exists by listing with prefix equal to that key and
only cares about the first result. We stat whether an object exists, and
if it does, we return it in a list with a continuation token set to this
key (signaling that there might be more). We additionally check for a
prefix object if the delimiter is a forward slash. If none was found, we
fall back to exhaustive listing.

3. Exhaustive listing

This is for arbitrary prefixes and delimiters. We list the entire
bucket, filter (based on prefix), and collapse paths on the gateway
side.

Since this change has grown significantly, only existing tests have been
adjusted (we now capture a wider range of cases), and extensions of the
test suite will be added in the next changesets.

Closes #43

Change-Id: I30df7be5d3ef3c9be8d7686604b8a02ee72917e0