Jenkinsfile,testsuite: add rclone tests
This is the same set up as uplink rclone tests except we are testing
the s3 backend of rclone instead of the native Storj DC integration.

Additionally, this disables the metainfo rate limiter for testing
Gateway-ST, just as Gateway-MT currently does. rclone tests throw
a lot of requests at the test Satellite, so we need to disable this.

Updates storj/gateway-mt#90

Change-Id: I54462f14a251db562375025cf5faabf22ba86673