{miniogw, testsuite/miniogw}: fix exhaustive ListObjects(V2)'s filtering
There are two issues fixed in this change:

First: one of the optimizations in the exhaustive listing had incorrect
logic. Second: keys weren't correctly filtered out when the continuation
token/start after string was a collapsed key (which happens in a
non-recursive listing).

For these to not happen again, this change adds additional tests that
stress described cases (specifically 2. case).

Change-Id: I0c23d4a7d329fbfbd90c5b575cd0311c9371248f