{miniogw, testsuite/miniogw}: rm checkBucketError from ListObjects(V2)
checkBucketError calls are no longer necessary in
ListObjects/ListObjectsV2 as all code paths lead to a listing that
reports bucket errors correctly. Previously, we needed them as we
stopped at the prefix-optimized listing, which uses StatObject under the
hood, which does not report the error correctly. Listing is now defined
as falling back to the exhaustive listing if nothing in the
prefix-optimized case was found, which means errors will be reported
correctly.

Change-Id: I8486d0ef033288fad2fb67f6016c9d8d8383be21