{docs,miniogw,testsuite/miniogw}: allow specifying object's TTL via meta
It's possible to specify TTL for an object with libuplink/uplink CLI,
and one way to achieve this with the gateway is to send a Storj-specific
header. Gateway now accepts MinIO-supported X-Amz-Meta-Storj-Expires and
X-Minio-Meta-Storj-Expires to do that. The value format is +Ab, where A
is integer and b is the time unit or full RFC3339-formatted date. The
change also adds relevant tests and updates documentation.

Updates storj/gateway-mt#195

Change-Id: I019e9d77993ae27321d75560ca1c1077242ecdf0