miniogw: S3 compatibility: fix ListMultipartUploads results
Make ListMultipartUploads work consistent with ListObjects(V2) in
determining the max uploads limit. This also fixes problems with
paging, as it was observed S3 CLI would not keep fetching more
results if you had over 1000 pending uploads in a bucket. It's
likely nobody encountered this problem yet because it's rare to
be listing large amounts of pending uploads.

Updates storj/gateway-mt#207

Change-Id: I172ba6ec3ff7739db284197de9ab1d0ff4891b6d