ci: rework integration tests to run mint and splunk tests too
This adds Makefile targets to run integration tests locally, similar
to how gateway-mt integration tests are run. This also means we'll
be running the same suite on changes to gateway-st as we do with
mt, so we can catch issues sooner.

Additionally, this changes GetBucketPolicy to return nil if
gateway is started without the `--website` flag, which fixes a
failing minio-go mint test that expected a nil policy. On gateway-mt
this is what we return as well, so it's more a fix for consistency.

Closes storj/gateway-mt#172

Change-Id: Ic0795b5cfc30c68b45ab4dd4e28f2ad6b0d15d1d