miniogw: make ListObjects/V2 aware of easier cases
This change changes listObjectsExhaustive (that ListObjects /
ListObjectsV2 is using) to not list too many items if it doesn't have
to. Specifically, if the prefix contains forward slashes, it will always
try to limit the scope of its listing by listing from the last forward
slash in the prefix.

Updates storj/customer-issues#184

Change-Id: Iae3fa0bc509cf089404a348bdaa9bae1c127966e