remove DialTimeout from config
In general, it's not recommended to remove a configuration option from
backwards compatible software. This is because users who rely on that
configuration option will be negatively impacted, and it could cause
compatibility issues for downstream users who are upgrading to the new
release.

However, current defaults for DialTimeout do more bad than good, and
setting this option does not affect certain connections. With this
context in mind, I decided to pull it out. Of course, the next release
will contain a comprehensive explanation.

Change-Id: I26705c8fc807da1f44c92bbddd5d4849da8c965b