Jenkinsfile: always clean up before and after integration tests
This change ensures that files created by and before the Integration
stage of the Jenkinsfile are always removed.

Previously, if Jenkins encountered an issue archiving test results at
the end of the Integration stage, the command to clean the integration
testing environment wouldn't be executed. This allowed files and Docker
containers to persist after the stage terminated.

Additionally, commit 1eaeff3 removed the code that cleaned up the
working directory from the beginning of the Integration stage because
it was assumed that the preceding Build stage would handle it once it
terminated. However, these stages aren't guaranteed to be run by the
same agent, so their working directories may be different. This commit
reintroduces the removed code.

Change-Id: I751fcbcfdff8e259c4d70a576468fd3e8ccef1c0