Makefile,Jenkinsfile: run ceph tests as current user
This change ensures that ceph integration tests are run as the user
that executed the command to start them rather than defaulting to root.
This ensures that files created in the host filesystem by these tests
can be deleted by that user.

Additionally, this change ensures that the latest revision of the
storj/up repository's main branch is used when installing storj-up.
This fixes an issue where the satellite container would reject project
creation requests originating from the old version of storj-up that we
used.

Change-Id: I5a3ba6ff657772874d3d075e3f27409a4fa0c65c