miniogw: fix up mappings to latest uplink code
- use the new uplink ErrRetentionNotFound error type. Additionally
this should return a 404 with the right status S3 uses. Note that
if the bucket has no lock enabled, GetObjectRetention does the same
as S3 and return a 400 response bucket lock not configured.

- uplink ErrBucketNoLock, ErrNoObjectLockConfiguration, and
ErrBucketNoVersioningObjectLock should all map to
miniogw.ErrRetentionNotFound, a 400 status. We will need to also
map ObjectLockDisabled (for when the feature is disabled entirely)
as well as ProjectLockDisabled to this same error, but these will
come in a later change.

Updates storj/edge#498

Change-Id: Ic3702e24cb3e8f177eac303ccb3f9869a8de96b5