miniogw: properly identify old OL disabled error for GetObjectRetention
This change properly identifies and returns the correct HTTP status
code of errors that occur when using GetObjectRetention on an object in
a bucket without Object Lock enabled. This was done in anticipation of
a libuplink dependency upgrade that changes the way that errors are
identified by libuplink.

Our build system runs integration tests using an old version of the
satellite that returns certain errors in a different format than the
latest version of libuplink expects them to be in. One of these errors
is that which is returned when a GetObjectRetention operation is
performed on an object in a bucket without Object Lock enabled. Because
of this, when our libuplink dependency version is bumped, we will not
be able to rely on libuplink to identify this error; we must do so
ourselves.

Change-Id: Iae4117f52569d9e6c29476a56608275525aa9523