{miniogw, testsuite}: rename --s3.upload-part-copy.disable-globally
After thinking about this, DisableCopyObject being false by default
makes sense because it's a metadata-only operation, and the availability
of the function is ultimately controlled by the satellite. This is not
the case for the gateway-side UploadPartCopy, so I believe changing it
to be disabled by default is more intuitive.

Change-Id: If384feb2902468153f141b9094452c8c6a936172