{Makefile,testsuite}: fix build from ci image bump
ci image now uses Debian trixie which has a Python and tox version
too new for ceph/s3-tests, so we must downgrade. This changes the
test to run on python:3.11-bookworm image and first installs tox
as part of the test. This is relatively fast, so shouldn't present
a problem in delay of tests.

This also removes a Duplicity parameter that no longer exists.

note: umask 0000 ensures all files can be cleaned up, because we
need to run the container as root to install tox.

Change-Id: I4875edab5d96e0e0fc9bf7e0e3d6d4c970cb74cf